CORE-575: rspace.history: restructure a bit
