+++ rspace/src/main/scala/coop/rchain/rspace/history/ITrieStore.scala
@@ -1,5 +1,12 @@
 package coop.rchain.rspace.history
 
+import cats.instances.option._
+import cats.instances.vector._
+import cats.syntax.traverse._
+
+import scala.annotation.tailrec
+import scala.collection.immutable.Seq
+
 trait ITrieStore[T, K, V] {
 
   private[rspace] def createTxnRead(): T
@@ -18,7 +25,35 @@ trait ITrieStore[T, K, V] {
 
   private[rspace] def toMap: Map[Blake2b256Hash, Trie[K, V]]
 
-  def close(): Unit
+  private[rspace] def getLeaves(txn: T, hash: Blake2b256Hash): Seq[Leaf[K, V]] = {
+    @tailrec
+    def loop(txn: T, ts: Seq[Trie[K, V]], ls: Seq[Leaf[K, V]]): Seq[Leaf[K, V]] =
+      ts match {
+        case Seq() =>
+          ls
+        case tries =>
+          val (next, acc) = tries.foldLeft((Seq.empty[Trie[K, V]], ls)) {
+            case ((nexts, leaves), Node(pointerBlock)) =>
+              val children =
+                pointerBlock.children
+                  .map(_._2)
+                  .traverse[Option, Trie[K, V]](hash => get(txn, hash))
+                  .getOrElse(throw new LookupException("something went wrong"))
+              (nexts ++ children, leaves)
+            case ((nexts, leaves), leaf: Leaf[K, V]) =>
+              (nexts, leaves :+ leaf)
+          }
+          loop(txn, next, acc)
+      }
+    get(txn, hash) match {
+      case Some(currentRoot) =>
+        loop(txn, Seq(currentRoot), Seq.empty[Leaf[K, V]])
+      case None =>
+        throw new LookupException(s"could not get node at $hash ")
+    }
+  }
 
   private[rspace] def clear(): Unit
+
+  def close(): Unit
 }
